data-science/notebooks/requirements.txt
azure-functions>=1.18.0
pandas
numpy
scikit-learn
matplotlib
seaborn
scipy
statsmodels
xgboost